Moves author profile tests
